requirements.txt
fastapi
uvicorn
pymongo>=4.3.3
python-dotenv
pydantic